Add CArray and CArrayType to CValue.
This adds both CArrayType and its CValue instance CArray to CValue. To
facilitate this a new sub-type of CType was added; CValueType[A]. Also,
CValue got CWrappedValue[A] for instances of these types. The hope is
that these will let us remove some boxing by being able to write generic
(specialized) methods with [@spec(Long, Double) A: CValueType]. It also
lets CArray wrap "unboxed" (by CValue) types. Once things have settled,
it would be worth it to specialize CArray on Boolean, Long, and Double
and take Array[A] instead of IndexedSeq[A].