Initial cut at n-ary Slice & Column concat.
This adds a new Column to ColumnSupport: NConcatColumn. It also adds a
new operation to cf.Util: NConcat(...). These are used by Slice's new
concat(slices) method that can concatenate several slices together and
work with them efficiently. The look-up time for a row is O(lg n) where
n is the number of concatenated columns. It fast-tracks the case where
the last row accessed is in the same column as the current row accessed.