Refactor use of @tailrec inside anonymous functions.
This commit finds uses of @tailrec affected by SI-6526 and rewrites
them at a higher scope, adding parameters as necessary. Note that
methods within methods are fine, it's only anonymous functions
which pose a problem. So for example:

  def outer(...) = {
    @tailrec def isOk(...) { ... }

    something map { xyz =>
      @tailrec def isNotOk(...) {...}
      ...
    }
  }

This commit shouldn't change any behavior. All the tests seem to be
passing. However I was necessarily reading the code out of, so review
(by anyone) would be appreciated.