Create a proper, lazy version of mapValues.
Map's mapValues doesn't actually create a map with the new values, but
rather wraps the old map and calls the conversion function every time an
element is accessed. This means we may be redoing a lot of work
unnecessarily because it runs the function multiple times. Just forcing
the map into existence isn't ideal either, as sometimes we're doing more
work than necessary because a column is never actually accessed. So, a
new method is added to Map, lazyMapValues, which can be treated like
mapValues, but will only call the function once per value (at most),
memoizing the result for subsequence calls.

In the athletes query, this provided a 5s speed up (~7.5%) on my
machine.