Update to blueeyes-1.0.0-M2, including new JSON parser.
This commit updates to a newer version of BlueEyes. The big JSON changes are:

  * types in JsonAST moved out to blueeyes.json
  * JsonParser -> JParser, with many changes to the parse methods
  * changes to JValue hierarchy
  * more JNum types and better numeric comparisons
  * JObject backed by Map, changes around ordering, pattern matching
  * removed Printers, added JValue#render* methods

In addition, there are other BlueEyes changes.

The new parser is less forgiving than the old: some test collateral was invalid
JSON and has been fixed. Code which was relying on the "looser" JSON parser
will need to be fixed (although all tests are passing).

There is still more work to do in terms of taking full advantage of the new
parser. This pull request is just about merging this branch to master to avoid
more drift.