Add sample method to Tables.
[Finishes #39284267]

This adds a new trait, SamplableTableModule, that adds a sample method
to Table. This let's you generate samples of a set size for the data in
a table. The current implementation is rather simple, requires a full
pass over the data, and runs in O(m^2 lg n + n) time.