Reimplemented findCommonality and associated machinery to use the appropriate graph traversal
This commit unifies the semantics of the GroupSolver in the compiler with the same functionality
in the evaluator.  Ideally, this functionality wouldn't exist in two places, but we're stuck with
it for the present due to the fact that we have orthogonal representations with insufficient
metadata transfer between them.  Note that this commit resolves more bugs that I can count relating
to grouping and the infamous "O NOES" exception.  It does, however, introduce a slight regression
for certain queries involving solve constraints that must be rearranged and are pulled up through
a function call by the Solver where the function in question is also used in a non-solve context
with actuals that are not transpecable from the target set.  (I did say it was an edge case!)  The
compiler currently over-approximates this semantic and will reject some perfectly valid queries.

[Fixes #36808939][Fixes #39691939][Fixes #39590081][Fixes #38813037]