This commit updates platform to the newest version of BlueEyes.
Among other changes, Chunk and ByteChunk are now just type aliases:

    type Chunk[A] = Either[A , StreamT[Future, A]]
    type ByteChunk = Chunk[ByteBuffer]

There were also some changes to the service combinators. In general,
we're trying to move away from requiring "impossible bijections"
(for instance, those between A and Future[A]).

All tests are currently passing. This was a big learning exercise
for me, in terms of using StreamT, so I'm sure there are places
where things could be polished a bit. But I think this branch is
in good enough shape to be merged.