Provide HTTP error reporting in WebJobManager.
This changes WebJobManager's monad to use EitherT[Future, String, ?]
instead of just Future. The left side is used only to report HTTP
errors. This also adds a withM method to JobManager that can swap the
monad used in the JobManager (if a natural transformation exists). This
is used, for example, in testing to simplify a WebJobManager to just a
JobManager[Future].