Ensure freshIdScanner is restartable.
So, freshIdScanner always generates fresh IDs, everytime a slice is
accesed. However, it isn't restartable. So, if a table is passed over
twice, it'll generate different IDs for each table. This was discovered
investigating a bug where a Morph2 was passed the results of a solved
with reductions. The results were given a new ID, then Morph2 tried to
cogroup them, after constructing the left/right tables (both derefs) to
cogroup on. This meant both the left and right tables had different IDs,
given by the freshIdScanner, and so they could not be cogrouped. The
Morph2 function (correlation in this case) got a bunch of empty slices,
which was less than ideal.

[Finishes #38535135]