Managed queries can now be expired (almost).
This adds the ability for managed queries to expire gracefully. However,
the JobManager was also modified. Mainly, job managers should not be
expected to be able to expire a job themselves, so, now they can't! The
main missing piece is now passing an expiration into the executorFor
method on QueryExecutorFactory.