Add support for DB authentication to the mongoDB backend
[Fixes #39499783]

This allows the end-user to configure specific user/pass pairs for
particular databases. Unlisted databases won't use authentication.
Due to an elusive bug in BlueEyes, any exception thrown in the
load stream would hang the output, so we're temporarily emitting
a "None" to terminate the stream at that point instead.