Add data attr to jobs. Store query in job data.
This adds a new data (Option[JValue]) member to Job. This let's us store
some arbitrary information related to the job. Shard queries now also
use this to store the query itself in the job when a job is started. The
current query info stored is the query itself, the path prefix, the sort
fields (if any) and the sort order.