Add async queries to web service and add tests.
This adds the ability to create and access the results of an
asynchronous query. Posting a query to jobs/queries will create the
asynchronous query. This will return a JobId which can then be used to
access the query results. If the query completed successfully, then the
results are returned with 200 OK. If the query is still running, then
204 No Content is returned. If the query failed, was cancelled, or
expired, then a 410 Gone is returned (indicating that it is impossible
for the results to ever be created).

This only adds tests for the AsyncQueryExecutor itself, and not for the
shard service.