Get YggUtils using BlueEyes' new parseMany.
This change updates the test_data to use a newline-delimited format
instead of using a single JSON array. This means that eventually we
will be able to do asynchronous parsing. For now we're just using
parseManyFromFile.

All data has been converted and all tests are passing. This commit
also updates the BlueEyes dependency to 1.0.0-M6.3.