Refactor start-shard.sh to run services on random non-open ports to avoid conflicts with other builds
[Fixes #40810509]

In addition to the rework of start-shard.sh, tweaks were made to the
dev configs used by start-shard.sh to normalize and simplify the
sed-based replacements of port numbers in the configs. The run.sh
script was also updated to read the new port assignments from a common
file.

The random port assignment procedure in start-shard.sh does have a
small probability of dual-assigning a port because in some parts of
the script it preassigns several ports before starting the services,
but the risk is low compared to the additional complexity required in
the script to prevent this from happening.