Split findDescriptorRoot functionality out to better represent required ops
Now ensuring a root existence is separate from simply looking up the root (no
more createOk flag). This allows the archive op to be more efficient in the case
where the path to be archived does not, in fact, exist.

Also added more logging to ShardSystemActor and MetadataActor message routing
for debug.