Refactor ShardSystemActor to use per-component ActorSystems
[Fixes #42181297]

The issue we ran into here was that all Actors were using the default
ActorSystem, which has a bounded thread pool. During archives of large
numbers of columns, the ProjectionArchiveActor instances would block
in Await trying to find given projections' baseDirs, which is
accomplished via messages to the MetadataActor. This would essentially
live-lock the system because with all threads stuck in the Await
calls, there were no threads available for the Akka dispatcher to
process the MetadataActor's messages and respond to those Await
calls. Setting up separate ActorSystems for each Actor component
allows the MetadataActor and ProjectionsActor to each accomplish
needed work indepenent of the other components.