Add -Y option to check ingest consistency.
This command runs start-shard.sh twice, as follows:

  ( ./start-shard.sh -Z && ./start-shard.sh -R ) 2>&1 | grep ';;;'

The -Z invocation clear a temporary directory, starts the server and
begins an ingest of n100k.json (downloading from S3 if necessary).
After the ingest request returns, and as soon as count() returns a
non-zero result, the server is shut down forcefully (kill -9).

The -Y invocation loads from the same directory, and confirms that
eventually we end up with the correct number of rows (100k).

The output is filtered to make it easy to see if things work. It's
also easy to run the test by-hand:

  ./start-shard.sh -Z && ./start-shard.sh -R