Refactor of ShardService and query output formats.
This adds a new 'format option to synchronous queries that can be either
"simple" or "detailed" (default is "simple" if omitted). The simple
format just returns the data (query result) and nothing else. The
detailed format returns an object with "data", "warnings", and "errors"
fields.

This is also a substantial refactor of my previous refactoring of
ShardService. We're back to 1 ShardService trait. What services are
offered (ie. async queries, warnings/errors, etc.) depend on the
ShardState instead. Also, there is now only 1 required method,
configureShardState, and that is in charge of creating the various
requried services and returning either a BasicShardState or a
ManagedQueryShardState.

Lastly, the new "detailed" query output format should become the default
when we finally switch to a new version of the API.

PLATFORM-677 #finish