Update SBT to 0.12.2.
SBT 0.12.2 turned javaOptions into a Task instead of Setting, which
required a change. Besides that, the following plugins had their
revision updated:

* sbt-assembly 0.8.5
* sbt-idea 1.2.0
* ensime-sbt-cmd 0.1.0

Also disable parallel testing on mongo build, as the connection
pool is not set up for it. A better solution would be to set up
mongo so as to handle all the concurrent connections, but that
will probably take some fine tuning.