Disable staged rewrites inside of Splits
PLATFORM-678 #finish

Special casing SplitGroup and SplitParam was dangerously wrong. This
disables the special casing but also disables rewrites for children of
Split nodes.

We need to figure out a way to rewrite a child of a Split node. The
mapDown function only works when dealing with a whole tree and not a
subtree.