Get the shard service rendering CSV.
When the "Accept" header exists and contains text/csv the service will
now return CSV data. More precisely, the service looks for a MIME type
it can render (either application/json or text/csv), and uses the first
one it sees. If neither is present it defaults to application/json.