Allow rewriting of Split children
This passes through a sequence of Split parents. An initial Map in
mapDown is constructed from Split -> Split so that when SplitGroup and
SplitParam are rebuilt, it contains references to the old Split.

This is a bit of a hack - it only works if the new Split child is not
used in a wider view of the graph.

PLATFORM-875 #finish