Correctly handle mixed Long-, Double-, and NumColumns.
This commit makes sure to consider to compare numerical columns with the same
path but different CType. Previously, these commits would be broken. A test of
this is included. However, due to a bug in Table#renderJson, the test will only
use DecimalColumn by default. Changing Table#fromJson ended up being a bigger
undertaking, so I'm going to make that change separately (and later).

Currently we combine Long/Double/BigDecimal values into a single column. There
is probably a more principled solution, but given the sorting overhead of Rank,
this is pretty minor and simplifies the algorithm.