Improve error-handling in RawHandler.
This commit adds an error-handling test, fixes some bugs which the test
uncovered, and changes the RawHandler.load() method to provide more
information on whether the load was clean (no errors) or dirty (rawlog
corruption was detected).

Of course, a higher-level caller will need to check the event-ids loaded
anyway, but returning an additional Boolean status seems friendly.