Fix pandora test under sbt 0.12.2
For unknown reasons, testing pandora under sbt 0.12.2 produces SIGSEGV
faults. As a work-around, one can fork before testing, but that breaks
other things, described below, which also require fixing.

The extract-data task passes its directory as a system property, which
won't work with forked tests, so that has to be passed as a parameter
as well. Also, the forked JVM has -Xmx1g by default, which is not
enough, and no permgen configuration at all.