Add runtime sort info & make sorts memorable.
This adds runtime sort information to the evaluator, which can be used
to determine cases where we may have Cross{Left,Right}Sort with a
reduced table and others, and so our sorted order would still be valid.

Additionally, we also add a new optimization that pushes sorts down
the dag if their inner node is a Cross{Left,Right}Sort join with a
constant. This prevents repeated sorts from occuring, when the only
difference between them is a deref or a wrap.