Disable jdbc/mongo specs in normal build, improve their concurrency for packaging specs
PLATFORM-894 #finish
PLATFORM-764 #finish

Since they are the usual suspects for a number of build issues, better
to only run them when we're actually packaging their assemblies. For
the times we do need to run them, concurrency was improved by adding a
grace period between release and shutdown of their respective engines.