Improve performance of dates.
This commit changes the way we ingest dates. Previously we ingested all
dates as strings, and would parse them into dates as-needed (for instance
in TimeLib). This ended up being really slow. For ~1M rows, it was taking
<1s to read a string column, but 15s to read those values as dates.

With this change, ingest will use a simple heuristic to detect strings
that are possible dates. For values that match that heuristic which
also parse as dates, we will ingest them as DateTime objects into a
date column. This means that clients that happen to have ISO8601 dates
in string columns might "lose" those values. In the medium-term we will
be encouraging people to ingest "QSON" instead of JSON so it won't be
a problem.

Queries heavily using dates can expect major speed ups (7-12x) from this.