Add basic web service testing harness.
This commit adds an integration testing project ("gjallerhorn", which
was Heimdall's horn). Using Dispatch, it can hit a local (or remote)
shard to verify that the API calls are working correctly.

So far, we are successfully creating accounts, as well as testing some
basic security features. In the future, we should have tests for all
client-facing API services.