Fix security bug in NIHDB projections actor.
If we provided an API, but found a projections actor in the cache, we
were returning all the projections without checking if they have the
permissions to do so. However, we were also doing duplicate checks into
the cache, so just removing the cache check from findProjections fixes
this, as openProjection will check the cache anyways.