Adds -a to start-shard.sh
This option is related to -b. It takes a comma-separated list of
assembly targets (i.e. auth) and will force those assemblies to
be rebuilt no matter what. This is useful when one is hacking
on a module (i.e. shard) and testing with start-shard.sh. In this
case you'll want to rebuild the assembly every time you restart.

For example:

 ./start-shard.sh -b -a auth,accounts

This will build any missing assemblies (maybe -b should just become
the default? When do people not want this behavior?). It will also
build auth and accounts every time, even if they are up-to-date.