Big speed up to clustering.
Tom figured out how to tweak the clustering parameters a bit to do
a lot less work, and Erik did some low-level hackery on the various
cost calculations.

With a 400k row SITA-based data set we have gone from timing out on
a clustering query to completing it in 8.5s (mean of 6 runs after 3
warmup runs).