Expand string functions to support dates as well.
Since ingest will now opportunistically interpret strings as dates
(when using JSON) we need to be able to handle these dates when they
are treated as strings (since the user may not have intended them to
be used as dates).

Once we have smarter columns many of these hacks may disappear.

Fixes PLATFORM-1121