Add support for search-by-email and password reset on accounts
PLATFORM-1052 #finish

Adds handlers for account lookup by email address:

/accounts/search?email=<email address>

and password reset. Password reset is handled by calling

/accounts/<accountId>/password/reset

with a JSON body of {"email":"<account email address>"}. The email is verified
against the account Id for security. On success, a reset token and link is
sent to the account email, and the password is updated by POSTing to

/accounts/<accountId>/password/reset/<token>

with a  JSON body of {"password":"<new password>"}.

The link is hard-coded as

http://www.precog.com/reset-password-complete?service=<service name>&token=<token>&accountId=<accountId>

where <service name> is one of "beta.precog.com" or "nebula.precog.com".