Add serverWarnings channel for shard query jobs.
This adds a new ServerWarning channel to the jobs manager. This is used
to report problems with the server that won't prevent the execution of a
query. Currently, this is basically just when the jobs service itself is
down. It is similar to ServerError, in that the messages are usually
strings for queries, rather than line info.

PLATFORM-1217 #finish