Remove sort from Distinct in Evaluator.
There is no reason Table can't just run its own sort. Now that it does,
we also have the option to optimize it if necessary.

The case may be made that if the table is already sorted, then
re-sorting in Table#distinct is superfluous. However, we only use
Table#distinct once, in the Evaluator, and it sorts first without
question, so we aren't losing anything currently. In the future, we can
provide the ability to hint to distinct that the table is pre-sorted.
However, if the table is backed by a SQL table, we should let distinct
be handled by the SQB DB.