Be conservative about Table order after cross.
Our current cross implementation does not guarantee full identity order
after a cross. It only guarantees the left table's identities are in
order. We make note of this in the Evaluator now. In the future, we may
want to be able to optionally let cross know that we'd like to have
full-identity order if possible. It's a bit more work, but certainly
less work than having to join on the out-of-order identities.