Remove static sorting assumptions. Add uniqueness.
Since we no longer make use of the static sorting assumptions, we've
dropped them. We still maintain the list of value keys (ala ValueSort)
that the node has available at that point.

This also adds a new notion of static guaranteed identity uniqueness to
DepGraph nodes. We use this to determine the order of results out of a
cross, which can help save sorts of crossed tables in common cases.