Allow more flexibility in Evaluator crosses.
This allows the evaluator to provide hints to Table on how to perform
the cross. Cross may then decided to respect the hint or not. The idea
is to allow for optimizations that may require crosses to be performed
in a specific order to save a sort/join or 2 down the road.