Add std::string::editDistance function.
This commit adds a com.precog.util.Levenshtein object for calculating
edit distances (absolute and relative) as well as a function in
StringLib exposing this feature to Quirrel users.

The main tests for the Levenshtein distance itself are in util, and
a test to be sure the function is available is in pandora.

We aren't currently using normalized editDistance but I implemented it
in util since I figured we might need it eventually.