Allow Ops and Morphs to log and fail.
This replaces the use of EvaluationContext in Ops/Morphs with a new
MorphContext. Aside from wrapping EvaluationContext, this also provides
a logging capability similar to QueryLogger. The main difference is that
line location context is baked in, so the Op/Morph doesn't need to know
about it.

We can also use this to fail completely.